requirements.txt
fastapi
uvicorn
sqlalchemy
asyncpg
python-dotenv
mcp
google-genai